refactor: core mod and error handling
