+++ src/core/errors.rs
@@ -1,0 +1,16 @@
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum PullRequestError {
+    #[error("GitHub API error: {0}")]
+    ApiError(#[from] octocrab::Error),
+
+    #[error("Validation failed")]
+    PullRequestValidationFailed(String),
+
+    #[error("Invalid input: {0}")]
+    InvalidInput(String),
+
+    #[error("{0}")]
+    PATNotSet(String),
+}
