+++ src/core/input_mode.rs
@@ -1,0 +1,5 @@
+pub enum InputMode {
+    Normal,
+    Editing,
+    Creating,
+}
