+++ src/core/mod.rs
@@ -1,0 +1,4 @@
+pub mod app;
+pub mod errors;
+pub mod input_mode;
+pub mod pull_request;
